Refactor tenant related methods out in separate class - LOC is too damn high
